社区接口修改

+++ admin-console/src/main/java/io/information/modules/app/service/impl/InCardServiceImpl.java
@@ -1,9 +1,10 @@
 package io.information.modules.app.service.impl;
 
-import com.alibaba.fastjson.JSON;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.guansuo.common.StringUtil;
+import io.elasticsearch.entity.EsCardEntity;
+import io.information.common.utils.BeanHelper;
 import io.information.common.utils.IdGenerator;
 import io.information.common.utils.PageUtils;
 import io.information.common.utils.Query;
@@ -12,7 +13,6 @@ import io.information.modules.app.service.IInCardArgueService;
 import io.information.modules.app.service.IInCardBaseService;
 import io.information.modules.app.service.IInCardService;
 import io.information.modules.app.service.IInCardVoteService;
-import io.information.modules.app.vo.InCardVote;
 import io.mq.utils.Constants;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,28 +37,34 @@ public class InCardServiceImpl implements IInCardService {
     public void issueCard(InCard card, InUser user) {
         Long cId = IdGenerator.getId();
         InCardBase base = card.getBase();
+        EsCardEntity cardEntity = BeanHelper.copyProperties(base, EsCardEntity.class);
         if (null != card.getArgue() && StringUtil.isNotBlank(card.getArgue().getCaRside())
                 && null != card.getArgue().getCaFside() && StringUtil.isNotBlank(card.getArgue().getCaFside())) {
             InCardArgue argue = card.getArgue();
             argue.setcId(cId);
             argueService.save(argue);
-            rabbitTemplate.convertAndSend(Constants.cardExchange,
-                    Constants.card_Save_RouteKey, JSON.toJSON(argue));
+            if (null != cardEntity) {
+                cardEntity.setCaFside(argue.getCaFside());
+                cardEntity.setCaRside(argue.getCaRside());
+                cardEntity.setCaCloseTime(argue.getCaCloseTime());
+            }
         }
         if (null != card.getVote() && StringUtil.isNotBlank(card.getVote().getCvInfo())) {
             InCardVote vote = card.getVote();
             vote.setcId(cId);
             voteService.save(vote);
-            rabbitTemplate.convertAndSend(Constants.cardExchange,
-                    Constants.card_Save_RouteKey, JSON.toJSON(vote));
+            if (null != cardEntity) {
+                cardEntity.setCvCloseTime(vote.getCvCloseTime());
+            }
         }
         base.setcId(cId);
         base.setuId(user.getuId());
         base.setcCreateTime(new Date());
         baseService.save(base);
-        rabbitTemplate.convertAndSend(Constants.cardExchange,
-                Constants.card_Save_RouteKey, JSON.toJSON(base));
-
+        if (null != cardEntity) {
+            rabbitTemplate.convertAndSend(Constants.cardExchange,
+                    Constants.card_Save_RouteKey, cardEntity);
+        }
     }
 
     @Override
@@ -80,13 +86,11 @@ public class InCardServiceImpl implements IInCardService {
                     argue.setCaFsideNumber(fNumber.length);
                 }
                 card.setArgue(argue);
-                return card;
             }
             if (2 == category) {
                 //投票帖子
                 InCardVote vote = voteService.getById(base.getcId());
                 card.setVote(vote);
-                return card;
             }
         }
         //基础帖子
@@ -95,24 +99,26 @@ public class InCardServiceImpl implements IInCardService {
 
     @Override
     public void delete(Long[] cIds) {
-        for (Long cId : cIds) {
-            InCardBase base = baseService.getById(cId);
-            Integer category = base.getcCategory();
-            if (null != category) {
-                if (1 == category) {
-                    //辩论帖子
-                    argueService.removeById(cId);
+        if (null != cIds && cIds.length > 0) {
+            for (Long cId : cIds) {
+                InCardBase base = baseService.getById(cId);
+                Integer category = base.getcCategory();
+                if (null != category) {
+                    if (1 == category) {
+                        //辩论帖子
+                        argueService.removeById(cId);
+                    }
+                    if (2 == category) {
+                        //投票帖子
+                        voteService.removeById(cId);
+                    }
                 }
-                if (2 == category) {
-                    //投票帖子
-                    voteService.removeById(cId);
-                }
+                //基础帖子
+                baseService.removeById(cId);
             }
-            //基础帖子
-            baseService.removeById(cId);
+            rabbitTemplate.convertAndSend(Constants.cardExchange,
+                    Constants.card_Delete_RouteKey, cIds);
         }
-        rabbitTemplate.convertAndSend(Constants.cardExchange,
-                Constants.card_Delete_RouteKey, cIds);
     }
 
     @Override
@@ -129,5 +135,29 @@ public class InCardServiceImpl implements IInCardService {
                 queryWrapper
         );
         return new PageUtils(page);
+    }
+
+    @Override
+    public void update(InCard card) {
+        if (null != card) {
+            int category = card.getBase().getcCategory();
+            switch (category) {
+                case 1:
+                    InCardArgue argue = BeanHelper.copyProperties(card, InCardArgue.class);
+                    if (null != argue) {
+                        argueService.updateById(argue);
+                    }
+                    break;
+                case 2:
+                    InCardVote vote = BeanHelper.copyProperties(card, InCardVote.class);
+                    if (null != vote) {
+                        voteService.updateById(vote);
+                    }
+                    break;
+            }
+            baseService.updateById(card.getBase());
+            rabbitTemplate.convertAndSend(Constants.cardExchange,
+                    Constants.card_Update_RouteKey, card);
+        }
     }
 }
